screencap.sh: improved formatting, a bit of refactoring, and greatly enhanced error handling

+++ screencap.sh
@@ -10,45 +10,58 @@ usage: $(basename "$0") [-qr] [-d dir] [-f fmt]
 Creates and saves a screenshot using ImageMagick.
 Left click selects a window, left click and drag selects a region.
 
-Saves into XDG_PICTURES_DIR by default, creating it if it does not exist.
-If XDG_PICTURES_DIR is unset, ~/Pictures will be used.
-
 OPTIONS
-  -d dir:  Override the default directory. Will be created if needed.
-  -f fmt:  Override the default output format. Default is PNG.
+  -d dir:  Output directory. (Default: XDG_PICTURES_DIR or ~/Pictures if unset)
+           Missing directories will be created.
+  -f fmt:  Output format. (Default: png)
   -q:      Suppress notifications.
-  -r:      Screenshot the entire root window instead of a selection.
+  -r:      Select the root window.
 EOF
-  die -q "usage error"
+  err -fs $QUIET "usage error"
 }
 
-die() {
-  [ ! "$1" = "-q" ] && printf '%s\n' "$(basename "$0"): $1" 1>&2
-  [ "$1" = "-q" ] && shift
-  [ -z $NONOTIFY ] && notify-send -a "$(basename "$0")" -u critical -i dialog-error "Screencap failed" "$1"
-  kill -ABRT -- -$$
-}
+err() (
+  sev="alert"
+  unset quiet silent
+  while getopts fqs OPT; do
+    case $OPT in
+      f)  sev="failed" ;;  # fatal: cannot continue, abort execution
+      q)  quiet=1 ;;       # suppress notifications
+      s)  silent=1 ;;      # suppress stderr
+      ?)  ;;
+    esac
+  done
+  shift $((OPTIND - 1))
+
+  msg="${1:-Unknown error}"
+  app="$(basename "$0")"
+
+  [ -z $silent ] && printf '%s:%s\n' "$app" "$msg" 1>&2
+  [ -z $quiet ] && notify-send -a "$app" -u critical -i dialog-error "$app $sev" "$msg"
+  [ "$sev" = "failed" ] && kill -ABRT -- -$$
+  return 0
+)
 
 # shellcheck disable=SC2329
-abort() {
+_abort() {
   exit 1
 }
-trap 'abort' ABRT
+trap '_abort' ABRT
 
 depend() (
   for command in "$@"; do
-    type "$command" > /dev/null 2>&1 || die "FATAL ERROR: Required utility '$command' is missing."
+    type "$command" > /dev/null 2>&1 || err -f $QUIET "FATAL ERROR: Required utility '$command' is missing."
   done
 )
 
 checkdir() (
   for file in "$@"; do
     if [ -e "$file" ]; then
-      [ -d "$file" ] || die "$file exists but is not a directory."
-      [ -x "$file" ] && [ -w "$file" ] || die "$file exists but is not writable."
+      [ -d "$file" ] || err -f $QUIET "$file exists but is not a directory."
+      [ -x "$file" ] && [ -w "$file" ] || err -f $QUIET "$file exists but is not writable."
     else
       # shellcheck disable=SC2174
-      mkdir -p -m 0700 "$file" || die "$file doesn't exist and could not be created."
+      mkdir -p -m 0700 "$file" || err -f $QUIET "$file doesn't exist and could not be created."
     fi
   done
 )
@@ -71,19 +84,23 @@ capture() {
 }
 
 # shellcheck disable=SC2329
-clean() (
+_clean() (
   for file in "$@"; do
     [ -e "$file" ] && unlink "$file"
   done
 )
 
-unset CAPTURE_MODE CAPTURE_DIR CAPTURE_FMT NONOTIFY tempfiles
-type notify-send > /dev/null 2>&1 || NONOTIFY=1
+CAPTURE_DIR="${XDG_PICTURES_DIR:-"${HOME}/Pictures"}"
+CAPTURE_TMPDIR="${XDG_CACHE_HOME:-"${HOME}/.cache/$(basename "$0")"}"
+CAPTURE_FMT="png"
+
+unset CAPTURE_MODE QUIET
+type notify-send > /dev/null 2>&1 || QUIET=1
 while getopts d:f:qr OPT; do
   case $OPT in
     d)  CAPTURE_DIR="$OPTARG" ;;
     f)  CAPTURE_FMT="$OPTARG" ;;
-    q)  NONOTIFY=1 ;;
+    q)  QUIET="-q" ;;
     r)  CAPTURE_MODE="root" ;;
     ?)  usage ;;
   esac
@@ -91,24 +108,20 @@ done
 shift $((OPTIND - 1))
 
 depend magick
-[ ! "$WAYLAND_DISPLAY" ] && [ ! "$XDG_SESSION_TYPE" = "wayland" ] || die "ImageMagick import does not function on wayland."
-xset q > /dev/null 2>&1 || die "Can't find X session."
+[ ! "$WAYLAND_DISPLAY" ] && [ ! "$XDG_SESSION_TYPE" = "wayland" ] || err -f $QUIET "ImageMagick import does not function on wayland."
+xset q > /dev/null 2>&1 || err -f $QUIET "Can't find X session."
 
-[ "$CAPTURE_FMT" ] || CAPTURE_FMT="png"
-checkfmt "$CAPTURE_FMT" || die "output format '$CAPTURE_FMT' is not supported on your system."
-
-[ "$CAPTURE_DIR" ] || CAPTURE_DIR="${XDG_PICTURES_DIR:-"${HOME}/Pictures"}"
-CAPTURE_TMPDIR="${XDG_CACHE_HOME:-"${HOME}/.cache/$(basename "$0")"}"
+checkfmt "$CAPTURE_FMT" || err -f $QUIET "output format '$CAPTURE_FMT' is not supported on your system."
 checkdir "$CAPTURE_DIR" "$CAPTURE_TMPDIR"
-trap 'clean "$CAPTURE_TMPDIR/"*' EXIT
+trap '_clean "$CAPTURE_TMPDIR/"*' EXIT
 
-CAPTURE_TMP="$(_mkstemps "${CAPTURE_TMPDIR}/$(basename "$0")XXXXXX" ".${CAPTURE_FMT}")" || die "failed to create temp files"
-if [ -z $NONOTIFY ]; then
-  CAPTURE_ICON="$(_mkstemps "${CAPTURE_TMPDIR}/$(basename "$0")XXXXXX" ".jpg")" || die "failed to create temp files"
+CAPTURE_TMP="$(_mkstemps "${CAPTURE_TMPDIR}/$(basename "$0")XXXXXX" ".${CAPTURE_FMT}")" || err -f $QUIET "failed to create temp files"
+if [ -z $QUIET ]; then
+  CAPTURE_ICON="$(_mkstemps "${CAPTURE_TMPDIR}/$(basename "$0")XXXXXX" ".jpg")" || err -f $QUIET "failed to create temp files"
 fi
 
-capture "$CAPTURE_MODE" "$CAPTURE_TMP" || die "import failed for an unknown reason, most likely on wayland."
-[ -z $NONOTIFY ] && magick "$CAPTURE_TMP" -resize 128x128 "$CAPTURE_ICON"
+capture "$CAPTURE_MODE" "$CAPTURE_TMP" || err -f $QUIET "import failed for an unknown reason, most likely on wayland."
+[ -z $QUIET ] && magick "$CAPTURE_TMP" -resize 128x128 "$CAPTURE_ICON"
 
 CAPTURE_TIME="$(date +%F-%H%M%S)"
 CAPTURE_DIMS="$(magick identify -format '%wx%h' "$CAPTURE_TMP")"
@@ -116,8 +129,8 @@ CAPTURE_SUFX="screencap"
 
 CAPTURE_PATH="${CAPTURE_DIR}/${CAPTURE_TIME}_${CAPTURE_DIMS}_${CAPTURE_SUFX}.${CAPTURE_FMT}"
 
-link "$CAPTURE_TMP" "$CAPTURE_PATH" || die "rename failed"
-chmod "$(printf '%.4o\n' $((0666 & (~$(umask)))))" "$CAPTURE_PATH" || printf '%s\n' "$(basename "$0"): ALERT: failed to set file permissions"
-[ -z $NONOTIFY ] && notify-send -a "$(basename "$0")" -i "$CAPTURE_ICON" "Screenshot saved" "$CAPTURE_PATH"
+link "$CAPTURE_TMP" "$CAPTURE_PATH" || err -f $QUIET "rename failed"
+chmod "$(printf '%.4o' $((0666 & (~$(umask)))))" "$CAPTURE_PATH" || err $QUIET "ALERT: failed to set file permissions"
+[ -z $QUIET ] && notify-send -a "$(basename "$0")" -i "$CAPTURE_ICON" "Screenshot saved" "$CAPTURE_PATH"
 
 exit 0
